Replace local tracks with spotify versions
- Pretty print config to pm_settings.ini (was just string w/ no
formatting)
- Automatic (and optional) migration from local mp3s to spotify tracks
with (optional) removal of local file
- - currently tracks are migrated as they are played. Once the first
play finishes (to allow for thumbs down during that play) the file is
migrated. If the track or artist is not found, nothing happens and the
original file is left in place.
- - built remove.queue.add to use for the local track removal. Need to
ensure the files get removed but not until they are finished playing.
- built add.getTracks as generic handler for all spotify uris
(local,artist,album,track) and route to correct function to retrieve
object with all tracks matching the given criteria
- add.TracksFromAlbum.fromLocal - figures out the artist and gets all of
their albums, searches those albums for those that contain the track in
the URI (can be multiple albums) and adds all tracks from those albums
to the shuffle playlists.
- add.AlbumsFromArtist.fromLocal  - Figures out artist and gets all of
the tracks from all of their albums, and returns their uris in an array
- All remove.*.current functions now clear the deleteLaterTrack before
removing the file(s)